Vat-Value-Changed-To-Float

+++ databaseScripts/CreateDatabase.sql
@@ -181,7 +181,7 @@ GO
 
 CREATE TABLE [dbo].[VAT](
 	[VATID] [int] IDENTITY(1,1) NOT NULL,
-	[Value] [int] NOT NULL,
+	[Value] [float] NOT NULL,
 	[ActiveTimestamp] [datetime] NOT NULL, 
 	CONSTRAINT [PK_VAT] PRIMARY KEY CLUSTERED 
 	(
